subiendo cambios2

+++ app/Http/Controllers/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\models\Category;
+use Hamcrest\Type\IsObject;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -24,7 +25,21 @@ class CategoryController extends Controller
         //en esta funcion guardamos los datos enviados por post en la base de datos
         $data=$request->all();
         $category=Category::create($data);
-        return response()->json(['data'=> $category],200);
+        if(is_object($data)){
+          return response()->json([
+            'code'=>'200',
+            'status'=>'success',
+            'data'=> $category
+            ],200);
+           
+        }else{
+            return response()->json([
+                'status'=>'error',
+                'message'=>'la entrada no existe',
+                'code'=>'404'
+            ],404);
+
+        }
            
 
       
@@ -34,7 +49,20 @@ class CategoryController extends Controller
     public function show($id)
     {
         $category=Category::findorFail($id);
-        return response()->json(['data'=>$category],201);
+        if(is_object($category)){
+            return response()->json([
+                    'code'=>'200',
+                    'status'=>'success',
+                    'data'=>$category
+            ],201);
+        }else{
+            return response()->json([
+                        'status'=>'error',
+                        'message'=>'la entrada no existe',
+                        'code'=>'404'
+            ],404);
+        }
+        
     }
 
     
@@ -46,17 +74,61 @@ class CategoryController extends Controller
         ]);
         //realizamos la actualizacion del los datos
         $category=Category::findorFail($id);
-        $category->name=$request->name;
-        $category->save();
-        return  response()->json(['data'=>$category],200);
+        if(is_object($category)){
+            $category->name=$request->name;
+           if($category->save()){
+                return  response()->json([
+                    'data'=>$category,
+                    'status'=>'success',
+                    'code'=>'200'
+            
+                ],200);
+            }else{
+                return  response()->json([
+                    'message'=>'error al actualizar',
+                    'status'=>'error',
+                    'code'=>'404'
+                
+                ],200);  
+            }
+        
+        }else{
+            return response()->json([
+                'status'=>'error',
+                'message'=>'la entrada no existe',
+                'code'=>'404'
+            ],404);
+        }
+       
     }
 
   
     public function destroy($id)
     {
         $category=Category::findorfail($id);
-       if( $category->delete()){
-           return response()->json(['data'=>$category]);
-       }
+        if(is_object($category)){
+            
+            if( $category->delete()){
+                return response()->json([
+                    'code'=>'200',
+                    'status'=>'success',
+                    'data'=>$category
+                 ]);
+            }else{
+                return response()->json([
+                    'code'=>'404',
+                    'status'=>'error',
+                    'message'=>'error al eliminar'
+                ],404);
+
+            }
+        }else{
+            return response()->json([
+                'status'=>'error',
+                'message'=>'la entrada no existe',
+                'code'=>'404'
+            ],404);
+        }
+     
     }
 }
